catalog refactor

+++ NetMicroservice.Catalog.Api/Features/Courses/Feature.cs
@@ -3,8 +3,7 @@ namespace NetMicroservice.Catalog.Api.Features.Courses;
 public class Feature
 {
     public int Duration { get; set; }
-
-    public int Rating { get; set; }
+    public float Rating { get; set; }
 
     public string EducatorFullName { get; set; } = default!;
 }
